qna_module/requirements.txt
httplib2==0.9.2
matplotlib==2.2.2
numpy==1.15.4
pandas==0.23.4
PyYAML==3.12
scikit-learn==0.20.3
scipy==1.1.0
seaborn==0.9.0
urllib3==1.24.1